Add the Sign In activity
Add the colours to the colors.xml
Add the dimensions to the dimens.xml
Add all of the strings used to the strings.xml

Created the view for both v-26 and normal Sign in layout.
Added the code to sign into an account through firebase.

Added permissions for the activities in the manifest.